Reference

Data Structures

UnitfulCoordinateSystems.AbstractCoordinate — Type

AbstractCoordinate{N}

Used to represent a Unitful N-dimensional coordinate.

UnitfulCoordinateSystems.CoordinateRectangular — Type

CoordinateRectangular{L} <: AbstractCoordinate{2} where {L<:Unitful.Length}

Used to represent a two-dimensional rectangular coordinate located on the xy-plane and defined by orthogonal x and y components.

UnitfulCoordinateSystems.CoordinatePolar — Type

CoordinatePolar{L,A} <: AbstractCoordinate{2} where {L<:Unitful.Length, A<:DimensionfulAngles.Angle}

Used to represent a two-dimensional polar coordinate located on the xy-plane and defined by: a radius r, and an azimuth angle phi measured relative to the positive x-axis.

UnitfulCoordinateSystems.CoordinateCartesian — Type

CoordinateCartesian{L} <: AbstractCoordinate{3} where {L<:Unitful.Length}

Used to represent a three-dimensional Cartesian coordinate, defined by orthogonal x, y, and z components.

UnitfulCoordinateSystems.CoordinateCylindrical — Type

CoordinateCylindrical{L,A} <: AbstractCoordinate{3} where {L<:Unitful.Length, A<:DimensionfulAngles.Angle}

Used to represent a three-dimensional cylindrical coordinate, defined by: a range rho from the origin to the point on the xy-plane subtended by the coordinate, an azimuth angle phi measured relative to the positive x-axis, and an orthogonal z component.

UnitfulCoordinateSystems.CoordinateSpherical — Type

CoordinateSpherical{L,A} <: AbstractCoordinate{3} where {L<:Unitful.Length, A<:DimensionfulAngles.Angle}

Used to represent a three-dimensional spherical coordinate, defined by: a radius r from the origin to the coordinate, a polar angle theta measured relative to the positive z-axis, and an azimuth angle phi measured relative to the positive x-axis.

Component Functions

These functions are not exported by default because they present a high probability of namespace collisions. Instead, they can be accessed by prepending the package name, i.e. UnitfulCoordinateSystems.x, or by importing them explicitly, e.g. using UnitfulCoordinateSystems: x, y, z.

UnitfulCoordinateSystems.x — Function

x(r̄::AbstractCoordinate)

Calculate the x̂-directed component of r̄, i.e. the inner product of x̂ and r̄.

UnitfulCoordinateSystems.y — Function

y(r̄::AbstractCoordinate)

Calculate the ŷ-directed component of r̄, i.e. the inner product of ŷ and r̄.

UnitfulCoordinateSystems.z — Function

z(r̄::AbstractCoordinate)

Calculate the ẑ-directed component of r̄, i.e. the inner product of ẑ and r̄.

UnitfulCoordinateSystems.r — Function

r(r̄::AbstractCoordinate)

Calculate the magnitude/radius of r̄, i.e. the inner product of r̂ and r̄.

UnitfulCoordinateSystems.rho — Function

rho(r̄::AbstractCoordinate)

Calculate the magnitude of the coordinate that r̄ subtends onto the xy-plane.

UnitfulCoordinateSystems.ρ — Function

ρ(r̄::AbstractCoordinate)

Calculate the magnitude of the coordinate that r̄ subtends onto the xy-plane.

UnitfulCoordinateSystems.phi — Function

phi(r̄::AbstractCoordinate)

Calculate the angle between the positive x-axis and the coordinate that r̄ subtends onto the xy-plane.

Note: phi, φ, and ϕ are all aliased to the same function.

UnitfulCoordinateSystems.φ — Function

φ(r̄::AbstractCoordinate)

Calculate the angle between the positive x-axis and the coordinate that r̄ subtends onto the xy-plane.

Note: phi, φ, and ϕ are all aliased to the same function.

UnitfulCoordinateSystems.ϕ — Function

ϕ(r̄::AbstractCoordinate)

Calculate the angle between the positive x-axis and the coordinate that r̄ subtends onto the xy-plane.

Note: phi, φ, and ϕ are all aliased to the same function.

UnitfulCoordinateSystems.theta — Function

theta(r̄::AbstractCoordinate)

Calculate the angle between the positive z-axis and r̄.

Note: theta, θ, and ϑ are all aliased to the same function.

UnitfulCoordinateSystems.θ — Function

θ(r̄::AbstractCoordinate)

Calculate the angle between the positive z-axis and r̄.

Note: theta, θ, and ϑ are all aliased to the same function.

UnitfulCoordinateSystems.ϑ — Function

ϑ(r̄::AbstractCoordinate)

Calculate the angle between the positive z-axis and r̄.

Note: theta, θ, and ϑ are all aliased to the same function.